Refine types for "ancestor" functions
The _get_ancestors() (previously _get_ancestor_list) now only returns
SuperStates, so it does not return the object itself in the list.  This
is OK, as we can act on the object in the process_event() method (which
we've done).  Hopefully this allows more type discipline in the future?

Also, I renamed a couple things to make them better match what they are,
for example common_parent, which was actually a common_ancestor.