Test (informally) the 'heirarchical' in HSM
This is just an informal test, to make me feel better.  It gets me close
to a unit test, but still allows me to live dangerously by not having a
proper unit test.

(Seriously though, I do need to implement proper unit tests.)

I also added a warning when "upgrading" a State to a SuperState.  I
might add a strict mode that disallows such nonsense, and a "live
dangerously" mode for when I want to goof off in iPython or something,
that allows me to do the sorts of things that aren't best practices.

And remeber, if you're not living a life of danger, you may as well not
be living at all.  Live dangerously, my friends!