Add better type hinting for Callable actions
The `Callable` actions now include type hinting for the args (which is
empmty in all current cases) as well as the return types (which is None
in all current cases).

I'm doing this in preparation to add guards, which will (probably) be a
`Callable` which returns a `bool` (and I think takes nothing, but I'm
not quite sure yet).