Refactor and fix upgrading States to SuperStates
I had forgotten a parameter when I refactored previously.  Also, the
warning message should be given whenever a State is 'upgraded' as such,
and not just when it's done as the result of a sub-State being added to
it.

I also added tests for this, as I found myself testing it in IPython,
and figured if I'm going to do an impromptu test, I should just
formalize the test.