Change typing to be more backward compatible
When running `mypy` on this, it gave the following errors:

	stama/__init__.py:98: error: "dict" is not subscriptable, use "typing.Dict" instead  [misc]
	stama/__init__.py:98: error: X | Y syntax for unions requires Python 3.10  [syntax]

(I've currently got Python 3.8.10 on the machine I'm using.)

Since I want this library to be compatible with all currently supported
versions, I went ahead and made the suggested changes.  (Plus it was
fairly simple to do.)

On that note, I need to test this library against previous versions of
Python that are officially supported.  I need to figure out a good way
to automate that.