Implement Conditional Junctions
There were a few ways I was considering implementing this, but I think I
picked the right one.  The `ConditionalJunction` is a similar object to
a State (in fact I created a base class for both of them to use).  I
split the `StateMachine.process_event()` method into two methods:

1. `process_event()` which analyzes which state is the proper one to
   transition to.
2. And `transition_directly_to_event()` which handles the actually
   transition

(The first, `process_event()`, will run the second,
`transition_directly_to_event()` automatically.  Come to think of it,
I'm not so sure that `transition_directly_to_event()` shouldn't be a
private method, even if it will be called by `ConditionalJunction`.)

During `transition_directly_to_event()`, it checks to see if the next
state is a ConditionalJunction.  If so, ConditionalJunction checks its
logic and then runs `transition_directly_to_event()` directly.  (When
this happens, it's a nested/recursive function.)

Also, there are several bits of State that were replicated to
ConditionalJunction.  These should probably be placed into the base
class, to keep things DRY and avoid replication.

Finally, this all needs cleaning up a bit.  I didn't spend a whole lot
of time on things like docstrings, or do any linting, so I need to go
back and straighten up a bit.  (I was mostly focused on just getting it
working.)