Fix: entering starting state doesn't run actions
There was a bug, where on initially entering the starting state, none of
the appropriate actions would be run (on_entry for the state, as well as
any ancestor states).  The fix was initially simple: instead of simply
assigning the starting state to the current state, run the
transition_directly_to_state() method to transition into it as well.
The issue with this is that that method is not made to handle
transitioning to a state if there is no previous state, so it broke
things pretty quickly.  (How do you run the `on_exit` method on
`None`?)  Fixing that wasn't too hard;  I simply went through and didn't
run any methods on any previous state which was `None`.

All the tests passed, so it's working fairly well, but I don't quite
feel %100 percent good about this for some reason, and I can't figure
our why.  I don't have a good reason to _not_ move forward with it
either, so moving forward I am.