Add a lock on `process_event()`
So I realized that I may be a bit to lenient or lackadaisical with
threading in this library.  I had this realization after watching
Raymond Hettinger's "Keynote on Concurrency, PyBay 2017" (which is
excellent by the way -- you should check it out).  See:

https://www.youtube.com/watch?v=9zinZmE3Ogk

Anyway, I changed `process_event()` to add a lock on the state machine,
so that even if `transition_directly_to_state()` is called while
`process_event()` is running, it shouldn't mess anything up.

I mostly considered the above scenario (when two different processes
want to change the state of the state machine at the same or overlapping
times).  I failed, however, to consider what happens when someone (or
something) is trying to change a state or event when a state machine is
already transitioning, say like modifying the transitions or actions.
Icky race conditions!  I need to go through and analyze such things.  (I
don't want to go adding locks to everything willy-nilly, even though I'm
going to have to add a decent number of locks eventually.  In other
words, I want to do so conscientiously.)

In the mean time, I need to make sure to operate in two phases:  first,
setting up my state machine, and then second (only when the first phase
is completely done) running the state machine.  The two should not
overlap (at least until I get the locking situation refined).  The good
news: this is how I usually operate anyway (and how this would be used
in most situations), so it shouldn't be a big deal for the short term.

TODO: Add locks in the appropriate places.

(Bleh!)